Npm updates (#58)
* Don't use lodash chain

* Update alle the tings

A few warnings decided to show up.

* chunk commons [mini-css-extract-plugin]Conflicting order
* invalid prop `className` supplied to `React.Fragment`

The latter seems to come out of react-awesome-reveal which I had
to use as a drop-in replacement for react-reveal due to it no
longer bein maintained and therefore not being friends with newer
versions of React.

The former... the former seems to come from various modules being
imported in different order around in the source. I tried to sort
the imports by module in IntelliJ but it did not take and the
next step might be to go through every single import in every
single component and ensure things are imported in a consistent order?
I can't be arsed. It seems to work just fine.

* Update node for pipeline